pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.2.1",
    "numpy>=2.2.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.3.0",
    "scikit-learn>=1.6.1",
    "streamlit>=1.41.1",
    "trafilatura>=2.0.0",
]
